Fix even/odd intersection logic
Instead of checking if the previous segment does not span the slice, check
if the intersection point is the same as the previous point to determine
if we should not flip even/odd.

This probably won't work correctly for scenarios where a polygon has a
segment running along the slice.  It does fix a problem however where
a wrong turn is made and polygon is not sliced correctly.