ModelGenerator - fix problem where slices went missing
Instead of multipling n * layerDist, using addition to increment the
coordinates of the line segment, fixes the problem where slices go
missing.  I believe with the multiplication technique there was a rounding
error in the code.

Also, when --segment-debug is enabled, only print the before segmentation
images, and draw a grid first showing where the segmentation lines are.

Also, play around with how the faces are colored, see getColorForFace.